# api/swagger.yaml
definitions:
  github_com_SiriusServiceDesk_application-service_internal_models.Priority:
    enum:
    - Низкий
    - Средний
    - Высокий
    type: string
    x-enum-varnames:
    - Low
    - Medium
    - High
  github_com_SiriusServiceDesk_application-service_internal_models.Status:
    enum:
    - В работе
    - Создана
    - Выполнена
    - Отклонена
    type: string
    x-enum-varnames:
    - InProgress
    - Pending
    - Executed
    - Canceled
  github_com_SiriusServiceDesk_application-service_internal_web.CreateApplicationRequest:
    properties:
      comment:
        example: любой коммент или пустота
        type: string
      performer:
        example: Методический отдел
        type: string
      title:
        example: Любой заголовок до 20 символов
        type: string
    type: object
  github_com_SiriusServiceDesk_application-service_internal_web.GetApplicationResponse:
    properties:
      applicant:
        example: 23ger34-khsdb23G-23afh75-sdHvd
        type: string
      comment:
        example: любой коментарий ваще
        type: string
      create_date:
        example: 21.05.2024
        type: string
      execution_period:
        example: 7 рабочих дней
        type: string
      feedback:
        example: тут инфа видно только админу
        type: string
      id:
        example: "000000001"
        type: string
      performer:
        example: Методический отдел
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_models.Priority'
        example: низкий
      status:
        allOf:
        - $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_models.Status'
        example: Создана
      title:
        example: сломался кампутир
        type: string
    type: object
  github_com_SiriusServiceDesk_application-service_internal_web.GetApplicationResponseDoc:
    properties:
      details: {}
      payload:
        properties:
          applicant:
            example: 23ger34-khsdb23G-23afh75-sdHvd
            type: string
          comment:
            example: любой коментарий ваще
            type: string
          create_date:
            example: 21.05.2024
            type: string
          execution_period:
            example: 7 рабочих дней
            type: string
          feedback:
            example: тут инфа видно только админу
            type: string
          id:
            example: "000000001"
            type: string
          performer:
            example: Методический отдел
            type: string
          priority:
            allOf:
            - $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_models.Priority'
            example: низкий
          status:
            allOf:
            - $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_models.Status'
            example: Создана
          title:
            example: сломался кампутир
            type: string
        type: object
      status:
        example: 200
        type: integer
    type: object
  github_com_SiriusServiceDesk_application-service_internal_web.GetApplicationsResponseDoc:
    properties:
      details: {}
      payload:
        properties:
          applications:
            items:
              $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_web.GetApplicationResponse'
            type: array
        type: object
      status:
        example: 200
        type: integer
    type: object
  github_com_SiriusServiceDesk_application-service_internal_web.UpdateApplicationRequest:
    properties:
      execution_period:
        example: 7 рабочих дней
        type: string
      feedback:
        example: причина отказа или комментарий админа
        type: string
      priority:
        allOf:
        - $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_models.Priority'
        example: Низкий
      status:
        allOf:
        - $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_models.Status'
        example: В работе
    type: object
  github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse:
    properties:
      details: {}
      payload: {}
      status:
        example: 200
        type: integer
    type: object
  internal_web_status.Response:
    properties:
      code:
        type: integer
      data:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/status/:
    get:
      description: Get the status of the API
      operationId: Status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_web_status.Response'
      summary: Get the status
  /v1/admin/applications/:
    get:
      consumes:
      - application/json
      description: Get all applications (Admin only)
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_web.GetApplicationsResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
      summary: Get Applications
      tags:
      - admin
  /v1/admin/applications/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing application by ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Update Application Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_web.UpdateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: application updated
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
      summary: Update Application
      tags:
      - applications
  /v1/applications:
    get:
      description: Получает список заявок для администратора или пользователя
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_web.GetApplicationsResponseDoc'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить заявки
      tags:
      - applications
    post:
      consumes:
      - application/json
      description: Создает новую заявку
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: Создание заявки
        in: body
        name: application
        required: true
        schema:
          $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_web.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
      security:
      - ApiKeyAuth: []
      summary: Создать заявку
      tags:
      - applications
  /v1/applications/{id}:
    get:
      description: Получает заявку по ID для администратора или пользователя
      parameters:
      - description: Bearer <token>
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_internal_web.GetApplicationResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_SiriusServiceDesk_application-service_pkg_response.RawResponse'
      summary: Получить заявку
      tags:
      - applications
swagger: "2.0"